update routing scheme
Sufia builds routes in a Rails-esque manner, by having models own
second-level urls:

		/generic_model/:id
		/collection/:id

This allows for items to be added/edited without belonging to a
collection. Our current workflow places collections at the root-
level and attaches works to them. This commit begins setting up
routes in that manner. Currently the routes are

		/:collectionId
		/:collectionId/works/:workId

We should be able to keep the Sufia schema relatively unchanged
and use the UI to set-up the collections business rules.