Update state/prop fields / warn user when exiting with unsaved changes
`mapStateToProps` fields have been updated to better align with
their counterparts in `state`. `selectedWork` and `selectedCollection`,
which mapped to `state.work.data` and `state.collection.data`
respectively, now reference their parent keys.

A flag has been added to `state.work` to track when a user makes
changes to a Work. When `state.work.changed` is true, the user
is presented with an alert warning that unsaved changes will
be lost.

This closes GitHub issue #1.