Store WorkEdit title copy in work-metadata state
This adds a field to the Work state which contains the work's
display title (`[${workId}] ${workTitle}`). Now when the title
is edited + not saved, the display title will not change (which
I found to indicate that the item was being autosaved).

TODO: need to make sure this is updated when the work data is saved.