update 'WorkEdit' flow to match 'CollectionEdit'
This adds an `updates` object to the `work` store object. Changes
made in the edit form are set in this object and are merged with
the original `data` object on save. With this approach, we won't
have to store certain values (namely, `title`) outside of the
data object in order for them to not change on the rerender
caused by data change.