add actions/reducers for collections + search terms
Collections are stored in an array at the top level of state. Currently,
we're only keeping display names and ids, though a case could be made for
brief descriptions + thumbnail urls in the future. Of course, these
will be scoped to the user's permissions via Sufia.

Search terms are stored as a vocabulary, using a term that is not
likely to be overwritten. This will need to be revisited as to
whether or not this is a good practice and/or if they should be
editable by Vocabulary managers.