Implement range facet slider (#22)
👍 

* flatten RangeLimit... data props

* Parse out facet-like options when building querystring for search

To make things easier on ourselves, anything that looks like a facet
is treated as a facet within the app (I'm looking at you range-limits!)
When the query string is constructed, however, facets with a `type`
property have their values added to the `options` object.

From what I can tell, the `blacklight_range_limit` is the only gem
we're using that will throw this wrench, so the fix is geared
primarily towards that, but should work with others.

Resolves #6.

* fix failing tests

* Add FacetDateRangeLimit component

RangeLimit only works with integers. This component allows ranges
to be created from timestamps + `<input type="month|date"/>`
elements. It still needs a wrapper component (similar to
`<RangeLimitFacetWrapper />`) that will parse ISO date values
(which, I believe, is how we'll be receiving/sending values from
the repository) into said timestamps.

This does a lot of computing + would be a good candidate for a
refactor. Perhaps using WebWorkers?

* clear up some file names

* begin splitting out common catalog functions

adds functions for range facets:
- calculateRange
- date-intervals (constants related to dates)
- formatDateValue

and adds tests for `calculateRange`

* add FacetRangeLimitDate component

* add RangeSliderDate tests

* Update webpack.config.js

Replaced EnvironmentPlugin method with DefinePlugin

* Create config.js

Used to pass Environment variables through this file instead of typing in terminal

* Replace config.js with env.config.js

* Changed config.js to env.config.js

* Finished implementing env.config.js

* Added close button for SeaDragonViewer and components that are used on Adam's draft wireframes

* add RangeSliderNumber tests

* add FacetRangeLimitNumber test

* add FacetRangeLimitDate and format-date-value tests

* add Facets for date ranges for testing

* add basic .editorconfig file

* begin FormField component

* add metadata form components (string, text, date)

- adds the following components:
	- StringInput
	- TextInput
- moves DateInput from src/components/form-elements to src/components/metadata
- removes tests from src/components/form-elements

* add WorkMetadataForm component

* add GenericWork schema type (which isn't very generic)

* add blank work value if last work value is removed

* Major update: Form updates + Vocabulary overhaul

Had a hard time splitting this into smaller commits, so it's all one
big bundle of changes.

- metadata form updates
	- add Select component
	- bug fixes + additional tests
- Vocabulary overhaul
	- change vocabulary management to use new MetadataForm components
	- update createVocabulary action to expect a Term object instead of
	  name/description key/vals
- remove almost all previous form components
	- still need to update ControlledVocabulary to new interface
- work form schemas
	- add schemas for GenericTerm and Vocabulary
	- update GenericWork

* #28 [WIP] Integrating a IIIF Image Information Response Object as a single tile source for a given Work resource for the OpenSeadragon viewer

* remove previous form-elements/Select component

* add ControlledVocabularyInput component to GenericWork schema

* #28 [WIP] Ensuring that the Image Reference Strip is rendered vertically for cases where a Work contains multiple FileSets

* add GenericVocabulary schema tests

* fix bug where facets aren't being toggled

* hide facet-groups that contain no items

* #28 Adding the "onClose" prop. value and propType; Inserting the Sass rule for transitioning between reference strip states; Integrating a (very minimal) test suite

* allow Facets to accept values of any propType

* flatten RangeLimit... data props

* Parse out facet-like options when building querystring for search

To make things easier on ourselves, anything that looks like a facet
is treated as a facet within the app (I'm looking at you range-limits!)
When the query string is constructed, however, facets with a `type`
property have their values added to the `options` object.

From what I can tell, the `blacklight_range_limit` is the only gem
we're using that will throw this wrench, so the fix is geared
primarily towards that, but should work with others.

Resolves #6.

* fix failing tests

* Add FacetDateRangeLimit component

RangeLimit only works with integers. This component allows ranges
to be created from timestamps + `<input type="month|date"/>`
elements. It still needs a wrapper component (similar to
`<RangeLimitFacetWrapper />`) that will parse ISO date values
(which, I believe, is how we'll be receiving/sending values from
the repository) into said timestamps.

This does a lot of computing + would be a good candidate for a
refactor. Perhaps using WebWorkers?

* clear up some file names

* begin splitting out common catalog functions

adds functions for range facets:
- calculateRange
- date-intervals (constants related to dates)
- formatDateValue

and adds tests for `calculateRange`

* add FacetRangeLimitDate component

* add RangeSliderDate tests

* add RangeSliderNumber tests

* add FacetRangeLimitNumber test

* add FacetRangeLimitDate and format-date-value tests

* add Facets for date ranges for testing

* ensure non-string facet values aren't duplicated in query

* parse range limit from querystring

* add range sliders to SearchWrapper

* fix failing lint