run e2e test suite from a single js file
When running the end-to-end tests, we need to have the following
processes running simultaneously:

1) a mock API server
2) a mock app server
3) the e2e test loop (generate bundle, run Nightwatch tests, clean
   up the generated files)

What makes this difficult to do is all three need to run concurrently.
This isn't something easily achieved with a single line bash script,
so I decided to have it handled with a Node script (`test/index.js`).
This allows us to boot both servers, run the test loop, and then kill
the first two processes when the test loop completes.

There's probably some Bash wizardry to do this, but I opted to go the
route of getting this out the door quicker.